refactor: dal and service according to the 3-L. source
