--- internal/services/service/order_service.go
@@ -1,17 +1,0 @@
-package service
-
-import (
-	"hot-coffee/internal/core/entities"
-)
-
-type OrderService interface {
-	CreateOrder(order entities.Order) error
-	GetOrders() ([]entities.Order, error)
-	GetOrder(id string) (entities.Order, error)
-	UpdateOrder(id string, order entities.Order) error
-	DeleteOrder(id string) error
-	CloseOrder(id string) error
-	GetTotalSales() (entities.TotalSales, error)
-	GetPopularMenuItems() ([]entities.MenuItemSales, error)
-	GetOpenOrders() ([]entities.Order, error)
-}
